Revert "BugID:23977317:porting cfs schedule on linux host"
This reverts commit 4e8bfa00c163ebf8dca3c0fa925f2a69295bdaf6.

Reason for revert: <INSERT REASONING HERE>

+++ platform/arch/linux/include/k_types.h
@@ -15,8 +15,8 @@
 #define RHINO_MM_FREE_DYE            0xABABABAB
 
 typedef uint32_t cpu_stack_t;
-typedef uint64_t hr_timer_t;
-typedef uint64_t lr_timer_t;
+typedef uint32_t hr_timer_t;
+typedef uint32_t lr_timer_t;
 typedef sigset_t cpu_cpsr_t;
 
 #endif /* K_TYPES_H */
